chore: refactor bson converter to its own package
