+++ crates/datasources/src/lib.rs
@@ -1,5 +1,6 @@
 //! Data source implementations.
 pub mod bigquery;
+pub mod bson;
 pub mod common;
 pub mod debug;
 pub mod excel;
